oct initial commit
oct is a tool for reading the rate of clocks in OMAP's hardware clock tree.
oct can also be used to change clock rates and other parameters such as
switching parent clocks dynamically.  From the README:

oct works by parsing the OMAP clock tree in debugfs via a depth-first search.
Once it finds the clock specified in the first argument then it will print the
relevant clock rate information, or oct will set the new rate for that clock
if a rate was specified.

Depending on other options it is also possible to print the rates for all of
the parent clocks up until the root clock, or even to print rate data for
every clock in the system with the --tree option.

By specifying two clock names and no rates oct can also be used to switch the
parent clock of a specified clock.

Any time that kernel operations fail will result in oct returning the value of
the last debugfs operation.  If clock names are invalid oct will return
-ENOENT. If parent/child clocks are not compatible then oct will return
-EINVAL.  If the clock rate is not writeable oct will return -EPERM.  All
failures result in verbose error messages sent to STDERR.

Signed-off-by: Mike Turquette <mturquette@ti.com>