main.c & Makefile: merge latest work
I haven't touched otc in a few months and want to save what I have so
far.  The result may not be very functional or correct, but I'm going to
format the drive that this code currently resides on, so pushing to my
github repo makes the most sense as a backup mechanism.

Signed-off-by: Mike Turquette <mturquette@ti.com>